Update repulsorlift engines
